Extract `ServiceInstancesPoolUpdater` basic abstraction
