+++ app/services/abstract_method_error.rb
@@ -1,0 +1,2 @@
+class AbstractMethodError < StandardError
+end
