+++ spec/service/service_instances_pool_updater_spec.rb
@@ -1,0 +1,8 @@
+require "rails_helper"
+
+RSpec.describe ServiceInstancesPoolUpdater do
+  subject(:service) { described_class.new }
+
+  it { expect(service).to respond_to(:call) }
+  it { expect { service.call }.to raise_error(AbstractMethodError) }
+end
